extract SolrDocumentBehavior::ModelWrapper to a generic location
this class's mock-ActiveModel behavior is useful outside of
`SolrDocumentBehavior`. extract it for reuse elsewhere.
